+++ packages/molecular/src/Tabs/Tabs.css
@@ -1,0 +1,27 @@
+.YnI-Tabs {
+  display: flex;
+  width: fit-content;
+
+  border-bottom: 1px solid rgb(128, 128, 128);
+
+  &.animation {
+    position: relative;
+
+    .YnI-Tabs-Indicator {
+      display: block;
+      position: absolute;
+      bottom: 0;
+      height: 2px;
+      background-color: rgb(0, 0, 0);
+      transition:
+        width 150ms,
+        left 150ms;
+    }
+  }
+
+  &:not(.animation) {
+    .YnI-Tab.active {
+      border-bottom: 2px solid rgb(0, 0, 0);
+    }
+  }
+}
