+++ packages/molecular/src/Tabs/index.ts
@@ -1,0 +1,2 @@
+export * from "./Tabs";
+export { default } from "./Tabs";
