+++ packages/molecular/src/index.ts
@@ -1,3 +1,4 @@
-export * from "./Toast";
-export * from "./Tab";
 export * from "./Grid";
+export * from "./Tab";
+export * from "./Tabs";
+export * from "./Toast";
